Move query to top-level query_legacy
